update the httponly cookie setup
