+++ routers/api_booking_delete.py
@@ -1,21 +1,20 @@
-from fastapi import APIRouter, Request, Depends, HTTPException
+from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
 import mysql.connector
 from utils.token_verify_creator import token_verifier
-from fastapi.security import OAuth2PasswordBearer
 
 router = APIRouter()
 headers = {"Content-Type": "application/json; charset=utf-8"}
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 @router.delete("/api/booking")
-async def api_booking_delete(request: Request,token: str = Depends(oauth2_scheme)):
+async def api_booking_delete(request: Request): # ,token: str = Depends(oauth2_scheme)
     try:
-        if token == 'null':
+        token = request.cookies.get("access_token")
+        if not token:
             content_data = {"error": True, "message": "Please log-in to access the booking page."}
             return JSONResponse(status_code=403,content=content_data, headers=headers)
+        
         token_output = token_verifier(token)
-
         if token_output:
             db_pool = request.state.db_pool.get("basic_db") 
             with db_pool.get_connection() as connection:
@@ -28,12 +27,6 @@ async def api_booking_delete(request: Request,token: str = Depends(oauth2_scheme
                         content_data = {"ok": True}
                         return JSONResponse(content=content_data, headers=headers)
                     
-    except HTTPException as http_err:
-        return JSONResponse(
-            status_code=http_err.status_code,
-            content={"error": True, "message": http_err.detail},
-            headers=http_err.headers
-        )
     except (mysql.connector.Error) as err:
         return JSONResponse(
             status_code=500,
