+++ routers/api_booking_get.py
@@ -1,23 +1,20 @@
-from fastapi import APIRouter, Request, Depends, Header
+from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
 from utils.token_verify_creator import token_verifier
-from fastapi.security import OAuth2PasswordBearer
 import mysql.connector
 
 router = APIRouter()
 headers = {"Content-Type": "application/json; charset=utf-8"}
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 @router.get("/api/booking")
-async def api_booking_get(request: Request,token: str = Depends(oauth2_scheme)):  # ,journey: str = Header(None)
-    # get 資料只能放在header，不能放在body
-    # 我之前在後端用post，可以把資料型態審核交給signup_data: SignUpDataModel，api那邊就不必多寫json.loads()，現在需要多寫這個識別機制
+async def api_booking_get(request: Request):  # ,token: str = Depends(oauth2_scheme)
     try:
-        if token == 'null':
+        token = request.cookies.get("access_token")
+        if not token:
             content_data = {"error": True, "message": "Please log-in to access the booking page."}
             return JSONResponse(status_code=403,content=content_data, headers=headers)
+        
         token_output = token_verifier(token)
-
         if token_output:
             db_pool = request.state.db_pool.get("basic_db") 
             with db_pool.get_connection() as connection:
@@ -48,3 +45,5 @@ async def api_booking_get(request: Request,token: str = Depends(oauth2_scheme)):
             content={"error": True, "message": str(err)},
             headers=headers
         )
+    # get 資料只能放在header，不能放在body
+    # 我之前在後端用post，可以把資料型態審核交給signup_data: SignUpDataModel，api那邊就不必多寫json.loads()，現在需要多寫這個識別機制
