+++ routers/api_user_post.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Request, Form
+from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
 import mysql.connector
 from pydantic import ValidationError
