+++ static/attraction.html
@@ -10,8 +10,8 @@
 
     <script src="/static/css_js_folder/js/pre_job.js" defer></script>
     <script src="/static/css_js_folder/js/attraction_script.js" defer></script>
-    <script src="/static/css_js_folder/js/signin_signout_section_TESTING.js" defer></script>
     <!-- <script src="/static/css_js_folder/js/signin_signout_section.js" defer></script> -->
+    <script src="/static/css_js_folder/js/signin_signout_section_TESTING.js" defer></script>
     <script src="/static/css_js_folder/js/before_booking.js" defer></script>
 
     <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/nprogress/0.2.0/nprogress.min.js"></script> -->
