+++ static/booking.html
@@ -12,7 +12,6 @@
     <script src="/static/css_js_folder/js/pre_job.js" defer></script>
     <!-- <script src="/static/css_js_folder/js/signin_signout_section.js" defer></script> -->
     <script src="/static/css_js_folder/js/signin_signout_section_TESTING.js" defer></script>
-
     <script src="/static/css_js_folder/js/before_booking.js" defer></script>
     <script src="/static/css_js_folder/js/after_booking.js" defer></script>
   </head>
